chore(kuma-cp) move discovery controller to runtime plugin
