+++ pkg/core/bootstrap/plugins.go
@@ -16,8 +16,6 @@ import (
 	_ "github.com/kumahq/kuma/pkg/plugins/config/k8s"
 	_ "github.com/kumahq/kuma/pkg/plugins/config/universal"
 
-	_ "github.com/kumahq/kuma/pkg/plugins/discovery/k8s"
-
 	_ "github.com/kumahq/kuma/pkg/plugins/runtime/k8s"
 
 	_ "github.com/kumahq/kuma/pkg/plugins/ca/builtin"
