+++ pkg/core/plugins/registry.go
@@ -11,7 +11,6 @@ const (
 	resourceStorePlugin pluginType = "resource-store"
 	secretStorePlugin   pluginType = "secret-store"
 	configStorePlugin   pluginType = "config-store"
-	discoveryPlugin     pluginType = "discovery"
 	runtimePlugin       pluginType = "runtime"
 	caPlugin            pluginType = "ca"
 )
@@ -33,7 +32,6 @@ type Registry interface {
 	ResourceStore(name PluginName) (ResourceStorePlugin, error)
 	SecretStore(name PluginName) (SecretStorePlugin, error)
 	ConfigStore(name PluginName) (ConfigStorePlugin, error)
-	Discovery(name PluginName) (DiscoveryPlugin, error)
 	RuntimePlugins() map[PluginName]RuntimePlugin
 	CaPlugins() map[PluginName]CaPlugin
 }
@@ -53,7 +51,6 @@ func NewRegistry() MutableRegistry {
 		resourceStore: make(map[PluginName]ResourceStorePlugin),
 		secretStore:   make(map[PluginName]SecretStorePlugin),
 		configStore:   make(map[PluginName]ConfigStorePlugin),
-		discovery:     make(map[PluginName]DiscoveryPlugin),
 		runtime:       make(map[PluginName]RuntimePlugin),
 		ca:            make(map[PluginName]CaPlugin),
 	}
@@ -66,7 +63,6 @@ type registry struct {
 	resourceStore map[PluginName]ResourceStorePlugin
 	secretStore   map[PluginName]SecretStorePlugin
 	configStore   map[PluginName]ConfigStorePlugin
-	discovery     map[PluginName]DiscoveryPlugin
 	runtime       map[PluginName]RuntimePlugin
 	ca            map[PluginName]CaPlugin
 }
@@ -92,14 +88,6 @@ func (r *registry) ConfigStore(name PluginName) (ConfigStorePlugin, error) {
 		return p, nil
 	} else {
 		return nil, noSuchPluginError(configStorePlugin, name)
-	}
-}
-
-func (r *registry) Discovery(name PluginName) (DiscoveryPlugin, error) {
-	if p, ok := r.discovery[name]; ok {
-		return p, nil
-	} else {
-		return nil, noSuchPluginError(discoveryPlugin, name)
 	}
 }
 
@@ -139,12 +127,6 @@ func (r *registry) Register(name PluginName, plugin Plugin) error {
 			return pluginAlreadyRegisteredError(configStorePlugin, name, old, csp)
 		}
 		r.configStore[name] = csp
-	}
-	if dp, ok := plugin.(DiscoveryPlugin); ok {
-		if old, exists := r.discovery[name]; exists {
-			return pluginAlreadyRegisteredError(discoveryPlugin, name, old, dp)
-		}
-		r.discovery[name] = dp
 	}
 	if rp, ok := plugin.(RuntimePlugin); ok {
 		if old, exists := r.runtime[name]; exists {
