+++ pkg/plugins/resources/k8s/native/go.mod
@@ -10,7 +10,7 @@ require (
 	github.com/onsi/gomega v1.9.0
 	github.com/pkg/errors v0.8.1
 	golang.org/x/net v0.0.0-20191004110552-13f9640d40b9
-	k8s.io/apimachinery v0.18.6
+	k8s.io/apimachinery v0.18.8
 	k8s.io/client-go v0.18.6
 	sigs.k8s.io/controller-runtime v0.6.0
 )
