+++ pkg/plugins/runtime/k8s/controllers/testdata/01.dataplane.yaml
