+++ pkg/plugins/runtime/k8s/controllers/testdata/01.pod.yaml
