+++ pkg/plugins/runtime/k8s/controllers/testdata/02.other-services.yaml
