+++ pkg/plugins/runtime/k8s/controllers/testdata/02.services-for-pod.yaml
