+++ pkg/plugins/runtime/k8s/controllers/testdata/03.dataplane.yaml
