+++ pkg/plugins/runtime/k8s/controllers/testdata/03.services-for-pod.yaml
