+++ pkg/plugins/runtime/k8s/controllers/testdata/04.dataplane.yaml
