+++ pkg/plugins/runtime/k8s/controllers/testdata/04.other-services.yaml
