+++ pkg/plugins/runtime/k8s/controllers/testdata/04.services-for-pod.yaml
