+++ pkg/plugins/runtime/k8s/controllers/testdata/05.dataplane.yaml
