+++ pkg/plugins/runtime/k8s/controllers/testdata/06.pod.yaml
