+++ pkg/plugins/runtime/k8s/controllers/testdata/06.services-for-pod.yaml
