+++ pkg/plugins/runtime/k8s/controllers/testdata/07.pod.yaml
