+++ pkg/plugins/runtime/k8s/controllers/testdata/07.services-for-pod.yaml
