+++ pkg/plugins/runtime/k8s/controllers/testdata/08.dataplane.yaml
