+++ pkg/plugins/runtime/k8s/controllers/testdata/08.services-for-pod.yaml
