+++ pkg/plugins/runtime/k8s/controllers/testdata/09.dataplane.yaml
