+++ pkg/plugins/runtime/k8s/controllers/testdata/09.services-for-pod.yaml
