+++ pkg/plugins/runtime/k8s/util/util_test.go
@@ -5,7 +5,7 @@ import (
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 
-	. "github.com/kumahq/kuma/pkg/plugins/discovery/k8s/util"
+	. "github.com/kumahq/kuma/pkg/plugins/runtime/k8s/util"
 
 	kube_core "k8s.io/api/core/v1"
 	kube_meta "k8s.io/apimachinery/pkg/apis/meta/v1"
